Add filters to control argolink templates.
This is for #6.

This should allow us to safely remove the `partials/argo-links.php`,
and make it so argolinks in search results for any theme appear
correctly.

Three filters for content are included.

1. `the_content` returns to DOM as stored in argolink options (or
default to our default)
2. `the_excerpt` returns the DOM `partials/content-argolinks.php`
generates in Largo.
3. `post_type_link` returns the source URL instead of a post
permalink. Thus search result listings will link from archive pages
directly to the remote site.